dev_requirements.txt
black==23.7.0
requests==2.28.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pyspark==3.2.0
pandas==2.0.2
delta-spark==1.1.0
pre-commit==3.3.3
pdoc==14.1.0
python-dotenv==1.0.0
pydoc-markdown==4.8.2